feat: implement & test worker.process_available_jobs()
This is the part that actually calls graphile_worker to run
jobs and do the things.

Document that we're just about production ready.
Document that the admin API is experimental and probably
you should write your own, I guess.
Missing features are waiting to see if my lifecycle hooks PR
to graphile_worker lands or not.