feat: return EnqueueOutcome instead of error for duplicate job keys
When queuing a job with a duplicate id, as you might do if you
are queuing jobs with deterministic ids, subsequent attempts will
fail, and add_job() will respond with a row not found error,
because it fails to find any newly-added job.

We now catch this error and handle it gracefully. The EnqueueOutcome
enum lets us distinguish between the case of this job being added
successfully, and it not being added because it was a duplicate.

pub enum EnqueueOutcome {
    // success gives you the job
    Enqueued(Box<Job>),
    // duplicate gives you the key back
    AlreadyInProgress { job_key: String },
}

This is arguably a bug in graphile_worker: it should use
fetch_optional(), not fetch_one(). OTOH, maybe it's intended.

This is a breaking change. Instead of `Result<Job, Error>`,
`enqueue()` now returns `Result<EnqueueOutcome, Error>`.
The upside is you can do idempotent job enqueueing patterns
where duplicate submissions are normal and not errors. This
happens to be the use case I wrote this library for, so yay.